Crate delegation

Delegation Pallet

Provides means of adding KILT delegations on chain and revoking them. Each delegation is based on a specific CType. The most basic delegation is just a root node to which you can add further delegations by appending them to the root node resulting in a tree structure.

Config

Call

Pallet

Terminology

Claimer:: A user which claims properties about themselves in the format of a CType. This could be a person which claims to have a valid driver’s license.

Attester:: An entity which checks a user’s claim and approves its validity. This could be a Citizens Registration Office which issues drivers licenses.

Verifier:: An entity which wants to check a user’s claim by checking the provided attestation.

CType:: CTypes are claim types. In everyday language, they are standardised structures for credentials. For example, a company may need a standard identification credential to identify workers that includes their full name, date of birth, access level and id number. Each of these are referred to as an attribute of a credential.

Attestation:: An approved or revoked user’s claim in the format of a CType.

Delegation:: An attestation which is not issued by the attester directly but via a (chain of) delegations which entitle the delegated attester. This could be an employee of a company which is authorized to sign documents for their superiors.

Assumptions

The maximum depth of a delegation tree is bounded by MaxParentChecks. This is not enforced when adding new delegations. However, you can only revoke up to MaxParentChecks many sub-delegations.

Re-exports

pub use crate::default_weights::WeightInfo;

pub use crate::delegation_hierarchy::*;

pub use crate::pallet::*;

Modules

benchmarking

default_weights

Autogenerated weights for delegation

delegation_hierarchy

migrations

mock

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

DelegationAc

Controls the access to attestations.