Type Alias delegation::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn create_hierarchy( origin: OriginFor<T>, root_node_id: DelegationNodeIdOf<T>, ctype_hash: CtypeHashOf<T> ) -> DispatchResult

Create a new delegation root associated with a given CType hash.

The new root will allow a new trust hierarchy to be created by adding children delegations to the root.

There must be no delegation with the same ID stored on chain, while there must be already a CType with the given hash stored in the CType pallet.

The dispatch origin must be split into

a submitter of type AccountId who is responsible for paying the transaction fee and

a DID subject of type DelegationEntityId who creates, owns and can revoke the delegation.

Requires the sender of the transaction to have a reservable balance of at least Deposit many tokens.

Emits RootCreated.

Weight: O(1)

Reads: [Origin Account], Roots, CTypes

Writes: Roots

pub fn add_delegation( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T>, parent_id: DelegationNodeIdOf<T>, delegate: DelegatorIdOf<T>, permissions: Permissions, delegate_signature: DelegateSignatureTypeOf<T> ) -> DispatchResult

Create a new delegation node.

The new delegation node represents a new trust hierarchy that considers the new node as its root. The owner of this node has full control over any of its direct and indirect descendants.

For the creation to succeed, the delegatee must provide a valid signature over the (blake256) hash of the creation operation details which include (in order) delegation id, root node id, parent id, and permissions of the new node.

There must be no delegation with the same id stored on chain. Furthermore, the referenced root and parent nodes must already be present on chain and contain the valid permissions and revocation status (i.e., not revoked).

The dispatch origin must be split into

a submitter of type AccountId who is responsible for paying the transaction fee and

a DID subject of type DelegationEntityId who creates, owns and can revoke the delegation.

Requires the sender of the transaction to have a reservable balance of at least Deposit many tokens.

Emits DelegationCreated.

Weight: O(1)

Reads: [Origin Account], Roots, Delegations

Writes: Delegations

pub fn revoke_delegation( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T>, max_parent_checks: u32, max_revocations: u32 ) -> DispatchResultWithPostInfo

Revoke a delegation node (potentially a root node) and all its children.

Does not refund the delegation back to the deposit owner as the node is still stored on chain. Requires to additionally call remove_delegation to unreserve the deposit.

Revoking a delegation node results in the trust hierarchy starting from the given node being revoked. Nevertheless, revocation starts from the leave nodes upwards, so if the operation ends prematurely because it runs out of gas, the delegation state would be consistent as no child would “survive” its parent. As a consequence, if the given node is revoked, the trust hierarchy with the node as root is to be considered revoked.

The dispatch origin must be split into

a submitter of type AccountId who is responsible for paying the transaction fee and

a DID subject of type DelegationEntityId who creates, owns and can revoke the delegation.

Emits C * DelegationRevoked.

Weight: O(C) where C is the number of children of the delegation node which is bounded by max_children.

Reads: [Origin Account], Roots, C * Delegations, C * Children.

Writes: Roots, C * Delegations

pub fn remove_delegation( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T>, max_removals: u32 ) -> DispatchResultWithPostInfo

Remove a delegation node (potentially a root node) and all its children.

Returns the delegation deposit to the deposit owner for each removed DelegationNode by unreserving it.

Removing a delegation node results in the trust hierarchy starting from the given node being removed. Nevertheless, removal starts from the leave nodes upwards, so if the operation ends prematurely because it runs out of gas, the delegation state would be consistent as no child would “survive” its parent. As a consequence, if the given node is removed, the trust hierarchy with the node as root is to be considered removed.

The dispatch origin must be split into

a submitter of type AccountId who is responsible for paying the transaction fee and

a DID subject of type DelegationEntityId who creates, owns and can revoke the delegation.

Emits C * DelegationRemoved.

Weight: O(C) where C is the number of children of the delegation node which is bounded by max_children.

Reads: [Origin Account], Roots, C * Delegations, C * Children.

Writes: Roots, 2 * C * Delegations

pub fn reclaim_deposit( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T>, max_removals: u32 ) -> DispatchResultWithPostInfo

Reclaim the deposit for a delegation node (potentially a root node), removing the node and all its children.

Returns the delegation deposit to the deposit owner for each removed DelegationNode by unreserving it.

Removing a delegation node results in the trust hierarchy starting from the given node being removed. Nevertheless, removal starts from the leave nodes upwards, so if the operation ends prematurely because it runs out of gas, the delegation state would be consistent as no child would “survive” its parent. As a consequence, if the given node is removed, the trust hierarchy with the node as root is to be considered removed.

The dispatch origin must be signed by the delegation deposit owner.

DepositReclaimed.

Weight: O(C) where C is the number of children of the delegation node which is bounded by max_removals.

Reads: [Origin Account], Roots, C * Delegations, C * Children.

Writes: Roots, 2 * C * Delegations

pub fn change_deposit_owner( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T> ) -> DispatchResult

Changes the deposit owner.

The balance that is reserved by the current deposit owner will be freed and balance of the new deposit owner will get reserved.

The subject of the call must be the owner of the delegation node. The sender of the call will be the new deposit owner.

pub fn update_deposit( origin: OriginFor<T>, delegation_id: DelegationNodeIdOf<T> ) -> DispatchResult

Updates the deposit amount to the current deposit rate.

The sender must be the deposit owner.

impl<T: Config> Pallet<T>

pub fn calculate_delegation_creation_hash( delegation_id: &DelegationNodeIdOf<T>, root_id: &DelegationNodeIdOf<T>, parent_id: &DelegationNodeIdOf<T>, permissions: &Permissions ) -> T::Hash

Calculate the hash of all values of a delegation creation transaction.

Weight: O(1)

pub fn is_delegating( identity: &DelegatorIdOf<T>, delegation: &DelegationNodeIdOf<T>, max_parent_checks: u32 ) -> Result<(bool, u32), DispatchError>

Check if an identity is the owner of the given delegation node or any node up the hierarchy, and if the delegation has not been yet revoked.

It checks whether the conditions are required for the given node, otherwise it goes up up to max_parent_checks nodes, including the root node, to check whether the given identity is a valid delegator of the given delegation.

Weight: O(P) where P is the number of steps required to verify that the dispatch Origin controls the delegation entitled to revoke the attestation. It is bounded by max_parent_checks.

Reads: Roots, P * Delegations

impl<T: Config> Pallet<T>

pub fn delegation_nodes<KArg>(k: KArg) -> Option<DelegationNodeOf<T>>where KArg: EncodeLike<DelegationNodeIdOf<T>>,

An auto-generated getter for DelegationNodes.

impl<T: Config> Pallet<T>

pub fn delegation_hierarchies<KArg>( k: KArg ) -> Option<DelegationHierarchyDetails<CtypeHashOf<T>>>where KArg: EncodeLike<DelegationNodeIdOf<T>>,

An auto-generated getter for DelegationHierarchies.